Revert to using QLabel() instead of QSpacerItem()
For some weird reason grid.setRowStretch() has no effect at all at the size of the row.
As an alternative Qt recommends using QSpacerItem to add spacing in a grid.
However, this uses fixed widths and heights. (Mis)using a QLabel seems a safer option with respect to differences between OS-ses and accessibility.